Refactoring/Sad Path Testing

+++ app/controllers/api/v1/users/squads_controller.rb
@@ -1,8 +1,12 @@
 class Api::V1::Users::SquadsController < ApplicationController
   def index
-    user = User.find_by(params[:id])
-    squads = UserSquad.where(user_id: user.id)
-    render json: UserSquadSerializer.new(squads), status: 201
+    user = User.find_by(id: params[:user_id])
+    if !user.nil?
+      squads = UserSquad.where(user_id: user.id)
+      render json: UserSquadSerializer.new(squads), status: 201
+    else
+      render json: { description: "Error: User not found" }, status: 404
+    end
   end
 
   def create
@@ -12,7 +16,12 @@ class Api::V1::Users::SquadsController < ApplicationController
   end
 
   def destroy
-    user_squad = UserSquad.find(params[:id])
-    user_squad.destroy
+    user_squad = UserSquad.find_by(id: params[:id])
+    if !user_squad.nil?
+      user_squad.destroy
+      render json: { description: "UserSquad deleted successfully" }, status: 204 
+    else
+      render json: { description: "Error: UserSquad not found" }, status: 404
+    end
   end
 end
